feat: Main menu
Present to the user a list of software to be installed and the current status of each software.

The user will have a menu to execute actions acording to the list above:

* Install
* Update
* Reinstall
* Uninstall
* Validate

In this first version, the menu wont have any action just the visual part, as application is been built this menus will start to execute actions to take the user to the desired funcionality.

closes #3